+++ docs/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v2.5.4
+## UNRELEASED
 
 ### Fix bugs